backend/requirements.txt
asyncpg==0.31.0
asyncio==4.0.0
redis==7.3.0
pymilvus==2.6.10
SQLAlchemy==2.0.48
alembic==1.18.4
aio-pika>=9.6.1
email-validator==2.3.0
python-dotenv==1.2.2
fastapi==0.135.1
uvicorn[standard]==0.42.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
fastapi-limiter==0.2.0
